// libs/web/components/src/molecules/inputs/checkbox/enum.ts
export enum CHECKBOX_SIZE {
  SM = "sm",
  MD = "md",
  LG = "lg",
}

export enum CHECKBOX_VARIANT {
  PRIMARY = "primary",
  ERROR = "error",
  WARNING = "warning",
}

export enum LABEL_SIZE {
  SM = "sm",
  MD = "md",
  LG = "lg",
}

export enum STATUS_MESSAGE {
  SUCCESS = "success",
  ERROR = "error",
  WARNING = "warning",
}